Module coreaudio::audio_unit::audio_format

Typification of the various AudioFormat codes and flags offered by the Core Audio API.

See the Core Audio Data Types Reference here for more info.

Reexports

pub use self::standard_flags::StandardFlags;

pub use self::linear_pcm_flags::LinearPCMFlags;

pub use self::apple_lossless_flags::AppleLosslessFlags;

pub use self::audio_time_stamp_flags::AudioTimeStampFlags;

Modules

apple_lossless_flags
A wrapper around the const AppleLosslessFlags.

audio_time_stamp_flags
A wrapper around the const AudioTimeStampFlags.

linear_pcm_flags
A wrapper around the const LinearPCMFlags.

standard_flags
A wrapper around the const StandardFlags.

Enums

AudioFormat
A type-safe representation of both the AudioFormatId and their associated flags.

Mpeg4ObjectId
"Used in the mFormatFlags field of an AudioStreamBasicDescription structure that describes an MPEG-4 audio stream to specify the type of MPEG-4 audio data.