coreaudio::audio_unit
Module macos_helpers

Structs

AliveListener

An AliveListener is used to get notified when a device is disconnected.

RateListener

Changing the sample rate is an asynchonous process. A RateListener can be used to get notified when the rate is changed.

Functions

audio_unit_from_device_id

Create an AudioUnit instance from a device id.

find_matching_physical_format

Find the closest match of the physical formats to the provided StreamFormat. This function will pick the first format it finds that supports the provided sample format, rate and number of channels. The provided format flags in the StreamFormat are ignored.

get_audio_device_ids

get_audio_device_ids_for_scope

List all audio device ids on the system.

get_audio_device_supports_scope

does this device support input / ouptut?

get_default_device_id

Helper function to get the device id of the default input or output device.

get_device_id_from_name

Find the device id for a device name.

get_device_name

Get the device name for a device id.

get_hogging_pid

Helper for hog mode (exclusive access). Get the pid of the process that currently owns exclusive access to a device. A pid value of -1 means no process owns exclusive access.

get_supported_physical_stream_formats

Get a vector with all supported physical formats as AudioBasicRangedDescriptions.

set_device_physical_stream_format

Change the physical stream format (sample rate and format) of a device.

set_device_sample_rate

Change the sample rate of a device. Adapted from CPAL.

toggle_hog_mode

Helper for hog mode (exclusive access). Toggle hog mode for a device. If no process owns exclusive access, then the calling process takes ownership. If the calling process already has ownership, this is released. If another process owns access, then nothing will happen. Returns the pid of the new owning process. A pid value of -1 means no process owns exclusive access.