Module coreaudio::audio_unit::render_callback

Reexports

pub use self::action_flags::ActionFlags;

pub use self::data::Data;

Modules

action_flags

data
Format specific render callback data.

Structs

Args
Arguments given to the render callback function.

InputProcFnWrapper
This type allows us to safely wrap a boxed RenderCallback to use within the input proc.

Type Definitions

InputProcFn
When set_render_callback is called, a closure of this type will be used to wrap the given render callback function.