Module coreaudio::audio_unit::render_callback::action_flags

Structs

ActionFlags

Handle
A safe handle around the AudioUnitRenderActionFlags pointer provided by the render callback.

Constants

DO_NOT_CHECK_RENDER_ARGS
If this flag is set, then checks that are done on the arguments provided to render are not performed. This can be useful to use to save computation time in situations where you are sure you are providing the correct arguments and structures to the various render calls.

OFFLINE_COMPLETE
This flag is set when an offline unit has completed either its preflight or performed render operation.

OFFLINE_PREFLIGHT
This is used with offline audio units (of type 'auol'). It is used when an offline unit is being preflighted, which is performed prior to when the actual offline rendering actions are performed. It is used for those cases where the offline process needs it (for example, with an offline unit that normalizes an audio file, it needs to see all of the audio data first before it can perform its normalization).

OFFLINE_RENDER
Once an offline unit has been successfully preflighted, it is then put into its render mode. This flag is set to indicate to the audio unit that it is now in that state and that it should perform processing on the input data.

OUTPUT_IS_SILENCE
This flag can be set in a render input callback (or in the audio unit's render operation itself) and is used to indicate that the render buffer contains only silence. It can then be used by the caller as a hint to whether the buffer needs to be processed or not.

POST_RENDER
Called on a render notification Proc, which is called either before or after the render operation of the audio unit. If this flag is set, the proc is being called after the render operation is completed.

POST_RENDER_ERROR
If this flag is set on the post-render call an error was returned by the audio unit's render operation. In this case, the error can be retrieved through the lastRenderError property and the audio data in ioData handed to the post-render notification will be invalid.

PRE_RENDER
Called on a render notification Proc, which is called either before or after the render operation of the audio unit. If this flag is set, the proc is being called before the render operation is performed.