Struct coreaudio::audio_unit::render_callback::action_flags::Handle

pub struct Handle {
    // some fields omitted
}

A safe handle around the AudioUnitRenderActionFlags pointer provided by the render callback.

This type lets a callback provide various hints to the audio unit.

For example: if there is no audio to process, we can insert the OUTPUT_IS_SILENCE flag to indicate to the audio unit that the buffer does not need to be processed.

Methods

impl Handle

fn get(&self) -> ActionFlags

Retrieve the current state of the ActionFlags.

fn bits(&self) -> u32

The raw value of the flags currently stored.

fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

fn is_all(&self) -> bool

Returns true if all flags are currently stored.

fn intersects(&self, other: ActionFlags) -> bool

Returns true if there are flags common to both self and other.

fn contains(&self, other: ActionFlags) -> bool

Returns true if all of the flags in other are contained within self.

fn insert(&mut self, other: ActionFlags)

Insert the specified flags in-place.

fn remove(&mut self, other: ActionFlags)

Remove the specified flags in-place.

fn toggle(&mut self, other: ActionFlags)

Toggles the specified flags in-place.

fn from_ptr(ptr: *mut AudioUnitRenderActionFlags) -> Self

Wrap the given pointer with a Handle.