Struct coreaudio::audio_unit::render_callback::data::NonInterleaved

pub struct NonInterleaved<S> {
    // some fields omitted
}

A wrapper around the pointer to the mBuffers array.

Methods

impl<S> NonInterleaved<S>

fn channels(&self) -> Channels<S>

An iterator yielding a reference to each channel in the array.

fn channels_mut(&mut self) -> ChannelsMut<S>

An iterator yielding a mutable reference to each channel in the array.