Struct coreaudio::audio_unit::render_callback::Args

pub struct Args<D> {
    pub data: D,
    pub time_stamp: AudioTimeStamp,
    pub bus_number: u32,
    pub num_frames: usize,
    pub flags: Handle,
}

Arguments given to the render callback function.

Fields
data: D

A type wrapping the the buffer that matches the expected audio format.

time_stamp: AudioTimeStamp

Timing information for the callback.

bus_number: u32

TODO

num_frames: usize

The number of frames in the buffer as usize for easier indexing.

flags: Handle

Flags for configuring audio unit rendering.

This parameter lets a callback provide various hints to the audio unit.

For example: if there is no audio to process, we can insert the OUTPUT_IS_SILENCE flag to indicate to the audio unit that the buffer does not need to be processed.