Module coreaudio::audio_unit::types

Core Audio's various const audio unit types identifiers represented as typesafe enums.

Oirginal documentation here.

Enums

EffectType
Effect (digital signal processing) audio unit subtypes for audio units provided by Apple.

FormatConverterType
Audio data format converter audio unit subtypes for AudioUnits provided by Apple.

GeneratorType
Audio units that serve as sound sources.

IOType
Input/output AudioUnit subtypes for AudioUnits provided by Apple.

MixerType
Audio mixing AudioUnit subtypes for AudioUnits provided by Apple.

MusicDeviceType
Audio units that can be played as musical instruments via MIDI control.

Type
Represents the different kinds of Audio Units that are available.