coreaudio::audio_unit::types
Enum EffectType

pub enum EffectType {

    PeakLimiter = 1_819_112_562,
    DynamicsProcessor = 1_684_237_680,
    LowPassFilter = 1_819_304_307,
    HighPassFilter = 1_752_195_443,
    BandPassFilter = 1_651_532_147,
    HighShelfFilter = 1_752_393_830,
    LowShelfFilter = 1_819_502_694,
    ParametricEQ = 1_886_217_585,
    Distortion = 1_684_632_436,
    Delay = 1_684_368_505,
    SampleDelay = 1_935_961_209,
    GraphicEQ = 1_735_550_321,
    MultiBandCompressor = 1_835_232_624,
    MatrixReverb = 1_836_213_622,
    Pitch = 1_953_329_268,
    AUFilter = 1_718_185_076,
    NetSend = 1_853_058_660,
    RogerBeep = 1_919_903_602,
    NBandEQ = 1_851_942_257,

}

Effect (digital signal processing) audio unit subtypes for audio units provided by Apple.

Variants

PeakLimiter = 1_819_112_562

An audio unit that enforces an upper dynamic limit on an audio signal.

Available in OS X v10.2 and later.

DynamicsProcessor = 1_684_237_680

An audio unit that provides dynamic compression or expansion.

Available in OS X v10.3 and later.

LowPassFilter = 1_819_304_307

An audio unit that passes frequencies below a specified cutoff frequency and blocks frequencies above that cutoff frequency.

Available in OS X v10.2 and later.

HighPassFilter = 1_752_195_443

An audio unit that passes frequencies above a specified cutoff frequency and blocks frequencies below that cutoff frequency.

Available in OS X v10.2 and later.

BandPassFilter = 1_651_532_147

An audio unit that passes frequencies between specified upper and lower cutoff frequencies, and blocks frequencies outside that band.

Available in OS X v10.2 and later.

HighShelfFilter = 1_752_393_830

An audio unit suitable for implementing a treble control in an audio playback or recording system.

Available in OS X v10.2 and later.

LowShelfFilter = 1_819_502_694

An audio unit suitable for implementing a bass control in an audio playback or recording system.

Available in OS X v10.2 and later.

ParametricEQ = 1_886_217_585

An audio unit that provides a filter whose center frequency, boost/cut level, and Q can be adjusted.

Available in OS X v10.2 and later.

Distortion = 1_684_632_436

An audio unit that provides a distortion effect.

Available in OS X v10.5 and later.

Delay = 1_684_368_505

An audio unit that introduces a time delay to a signal.

Available in OS X v10.2 and later.

SampleDelay = 1_935_961_209

An audio unit that provides a time delay for a specified number of samples.

Available in OS X v10.4 and later.

GraphicEQ = 1_735_550_321

An audio unit that provides a 10- or 31-band graphic equalizer.

Available in OS X v10.2 and later.

MultiBandCompressor = 1_835_232_624

An audio unit that provides four-bands of dynamic compression or expansion.

Available in OS X v10.3 and later.

MatrixReverb = 1_836_213_622

An audio unit that provides a reverberation effect that can be used to simulate a variety of acoustic spaces.

Available in OS X v10.2 and later.

Pitch = 1_953_329_268

An audio unit for modifying the pitch of a signal.

Available in OS X v10.4 and later.

AUFilter = 1_718_185_076

An audio unit that provides a combination of five filters: low-frequency, three mid-frequencies, and high-frequency.

Available in OS X v10.4 and later.

NetSend = 1_853_058_660

An audio unit for use in conjunction with a kAudioUnitSubType_NetReceive audio unit for sending audio across a network or from one application to another.

Available in OS X v10.4 and later.

RogerBeep = 1_919_903_602

An audio unit that detects gaps between segments of speech and fills the gaps with a short tone, simulating the sound of a walkie-talkie communication device.

Available in OS X v10.5 and later.

NBandEQ = 1_851_942_257

A multi-band equalizer with specifiable filter type for each band.

Available in OS X v10.9 and later.