Enum coreaudio::audio_unit::types::FormatConverterType

pub enum FormatConverterType {
    AUConverter,
    NewTimePitch,
    TimePitch,
    DeferredRenderer,
    Splitter,
    Merger,
    Varispeed,
    AUiPodTimeOther,
}

Audio data format converter audio unit subtypes for AudioUnits provided by Apple.

Variants
AUConverter

An audio unit that uses an audio converter to do linear PCM conversions, such as changes to sample rate, bit depth, or interleaving.

Available in OS X v10.2 and later.

NewTimePitch

An audio unit that can be used to have independent control of both playback rate and pitch.

In OS X it provides a generic view, so it can be used in both a UI and programmatic context.

It also comes in an offline version for processing audio files.

Available in OS X v10.7 and later.

TimePitch

An audio unit that can provide independent control of playback rate and pitch. This subtype provides a generic view, making it suitable for UI and programmatic context. OS X provides realtime and offline audio units of this subtype.

Available in OS X v10.3 and later.

DeferredRenderer

An audio unit that acquires audio input from a separate thread than the thread on which its render method is called.

You can use this subtype to introduce multiple threads into an audio unit processing graph.

There is a delay, equal to the buffer size, introduced between the audio input and output.

Available in OS X v10.4 and later.

Splitter

An audio unit with one input bus and two output buses. The audio unit duplicates the input signal to each of its two output buses.

Available in OS X v10.4 and later.

Merger

An audio unit with two input buses and one output bus. The audio unit merges the two input signals to the single output.

Available in OS X v10.4 and later.

Varispeed

An audio unit that can control playback rate. As the playback rate increases, so does pitch.

This subtype provides a generic view, making it suitable for UI and programmatic context.

OS X provides realtime and offline audio units of this subtype.

Available in OS X v10.3 and later.

AUiPodTimeOther

Available in OS X v10.9 and later.