Enum coreaudio::audio_unit::types::IOType

pub enum IOType {
    GenericOutput,
    HalOutput,
    DefaultOutput,
    SystemOutput,
    VoiceProcessingIO,
    RemoteIO,
}

Input/output AudioUnit subtypes for AudioUnits provided by Apple.

Variants
GenericOutput

An audio unit that responds to start/stop calls and provides basic services for converting to and from linear PCM formats.

Use this audio unit when sending the output of an audio processing graph to your application rather than to the output audio hardware. You would typically use the Generic Output unit for offline audio processing. Just like the other I/O units, the Generic Output unit incorporates a Format Converter unit. This lets the Generic Output unit perform format conversion between the stream format used in an audio processing graph and the format you want.

You can also use a Generic Output unit as the final node in a subgraph that you place into a parent audio processing graph.

Available in OS X v10.2 and later.

HalOutput

An audio unit that can provides input/output connection to an a specified audio device.

Bus 0 provides output to the audio device and bus 1 accepts input from the audio device.

Available in OS X v10.2 and later.

DefaultOutput

A specialized HalOutput audio unit that connects to the user’s selected default device in Sound Preferences.

Available in OS X v10.2 and later.

SystemOutput

A specialized HalOutput audio unit that connects to the user’s selected device for sound effects, alerts, and other user-interface sounds.

Available in OS X v10.2 and later.

VoiceProcessingIO

An audio unit that interfaces to the audio inputs and outputs of iPhone OS devices and provides voice processing features.

Bus 0 provides output to hardware and bus 1 accepts input from hardware.

See the Voice-Processing I/O Audio Unit Properties enumeration for the identifiers for this audio unit’s properties.

Available in OS X v10.7 and later.

RemoteIO

Connects to device hardware for input, output, or simultaneous input and output. Use it for playback, recording, or low-latency simultaneous input and output where echo cancelation is not needed.

See https://developer.apple.com/library/content/documentation/MusicAudio/Conceptual/AudioUnitHostingGuide_iOS/UsingSpecificAudioUnits/UsingSpecificAudioUnits.html Available in iOS.