Enum coreaudio::audio_unit::types::MixerType

pub enum MixerType {
    MultiChannelMixer,
    StereoMixer,
    Mixer3D,
    MatrixMixer,
}

Audio mixing AudioUnit subtypes for AudioUnits provided by Apple.

Variants
MultiChannelMixer

An audio unit that can have any number of input buses, with any number of channels on each input bus, and one output bus.

In OS X, the output bus can have any number of channels.

In iPhone OS, the output bus always has two channels.

Available in OS X v10.5 and later.

StereoMixer

An audio unit that can have any number of input buses, each of which is mono or stereo, and one stereo output bus.

Available in OS X v10.2 and later.

Mixer3D

An audio unit that can have any number of input buses and one output bus.

Each input bus can be mono, in which case it can be panned using 3D coordinates and parameters.

Stereo input buses pass directly through to the output.

Four-channel ambisonic inputs are rendered to the output configuration.

The single output bus can be configured with 2, 4, 5, 6, 7 or 8 channels.

Available in OS X v10.3 and later.

Deprecated in OS X v10.10.

MatrixMixer

An audio unit that can have any number of input and output buses with any number of channels on each bus.

You configure the mix using a matrix of channels with a separate input level control for each channel.

The audio unit also provides individual level control for each input-channel-to-output-channel combination, as well as level control for each output channel.

Finally, the audio unit provides a global level control for the matrix as a whole.

Available in OS X v10.3 and later.