Module coreaudio::audio_unit

This module is an attempt to provide a friendly, rust-esque interface to Apple's Audio Unit API.

Learn more about the Audio Unit API here and here.

TODO: The following are kAudioUnitSubTypes (along with their const u32) generated by rust-bindgen that we could not find any documentation on:

MIDISynth = 1836284270,

RoundTripAAC = 1918984547,

SpatialMixer = 862217581,

SphericalHeadPanner = 1936746610,

VectorPanner = 1986158963,

SoundFieldPanner = 1634558569,

HRTFPanner = 1752331366,

NetReceive = 1852990326,

If you can find documentation on these, please feel free to submit an issue or PR with the fixes!

Reexports

pub use self::audio_format::AudioFormat;

pub use self::sample_format::{SampleFormat, Sample};

pub use self::stream_format::StreamFormat;

pub use self::types::{Type, EffectType, FormatConverterType, GeneratorType, IOType, MixerType, MusicDeviceType};

Modules

audio_format
Typification of the various AudioFormat codes and flags offered by the Core Audio API.

render_callback

sample_format

stream_format
A rustification of the AudioStreamBasicDescription type.

types
Core Audio's various const audio unit types identifiers represented as typesafe enums.

Structs

AudioUnit
A rust representation of the au::AudioUnit, including a pointer to the current rendering callback.

Enums

Element
Represents the Input and Output Elements.

Scope
The input and output Scopes.