coreaudio::audio_unit::audio_format
Enum Mpeg4ObjectId

pub enum Mpeg4ObjectId {
    AAC_Main = 1,
    AAC_LC = 2,
    AAC_SSR = 3,
    AAC_LTP = 4,
    AAC_SBR = 5,
    AAC_Scalable = 6,
    TwinVQ = 7,
    CELP = 8,
    HVXC = 9,
}

“Used in the mFormatFlags field of an AudioStreamBasicDescription structure that describes an MPEG-4 audio stream to specify the type of MPEG-4 audio data.

Available in OS X v10.3 and later.

Deprecated in OS X v10.5.

Note: This type was originally represented using a bitflag field in the original API, however there is only ever one flag set at a time. Thus, we use an enum as a more accurate, user-friendly, type-safe representation.

Original documenation here.

Variants

AAC_Main = 1

Advanced audio coding; the baisc MPEG-4 technology.

AAC_LC = 2

Lossless coding; provides compression with no loss of quality.

AAC_SSR = 3

Scalable sampling rate; provides different sampling frequencies for different targets.

AAC_LTP = 4

Long term prediction; reduces redundancy in a coded signal.

AAC_SBR = 5

Spectral band replication; reconstructs high-frequency content from lower frequencies and side information.

AAC_Scalable = 6

Scalable lossless coding.

TwinVQ = 7

Transform-domain weighted interleaved vector quantization; an audio codec optimised for audio coding at ultra low bit rates around 8kbit/s.

CELP = 8

Code Excited Linear Prdiction; a narrow-band/wide-band speech codec.

HVXC = 9

Harmonic Vector Excitation Coding; a very-low bit-rate parametric speech codec.

Implementations

impl Mpeg4ObjectId

pub fn from_u32(u: u32) -> Option<Mpeg4ObjectId>

Create an Mpeg4ObjectId from a u32.