Enum coreaudio::audio_unit::audio_format::Mpeg4ObjectId

pub enum Mpeg4ObjectId {
    AAC_Main,
    AAC_LC,
    AAC_SSR,
    AAC_LTP,
    AAC_SBR,
    AAC_Scalable,
    TwinVQ,
    CELP,
    HVXC,
}

"Used in the mFormatFlags field of an AudioStreamBasicDescription structure that describes an MPEG-4 audio stream to specify the type of MPEG-4 audio data.

Available in OS X v10.3 and later.

Deprecated in OS X v10.5.

Note: This type was originally represented using a bitflag field in the original API, however there is only ever one flag set at a time. Thus, we use an enum as a more accurate, user-friendly, type-safe representation.

Original documenation here.

Variants
AAC_Main

Advanced audio coding; the baisc MPEG-4 technology.

AAC_LC

Lossless coding; provides compression with no loss of quality.

AAC_SSR

Scalable sampling rate; provides different sampling frequencies for different targets.

AAC_LTP

Long term prediction; reduces redundancy in a coded signal.

AAC_SBR

Spectral band replication; reconstructs high-frequency content from lower frequencies and side information.

AAC_Scalable

Scalable lossless coding.

TwinVQ

Transform-domain weighted interleaved vector quantization; an audio codec optimised for audio coding at ultra low bit rates around 8kbit/s.

CELP

Code Excited Linear Prdiction; a narrow-band/wide-band speech codec.

HVXC

Harmonic Vector Excitation Coding; a very-low bit-rate parametric speech codec.

Methods

impl Mpeg4ObjectId

fn from_u32(u: u32) -> Option<Mpeg4ObjectId>

Create an Mpeg4ObjectId from a u32.