coreaudio::audio_unit::render_callback::data
Trait Data

pub trait Data {
    // Required methods
    fn does_stream_format_match(stream_format: &StreamFormat) -> bool;
    unsafe fn from_input_proc_args(
        num_frames: u32,
        io_data: *mut AudioBufferList,
    ) -> Self;
}

Audio data wrappers specific to the AudioUnit’s AudioFormat.

Required Methods

fn does_stream_format_match(stream_format: &StreamFormat) -> bool

Check whether or not the stream format matches this type of data.

unsafe fn from_input_proc_args( num_frames: u32, io_data: *mut AudioBufferList, ) -> Self

We must be able to construct Self from arguments given to the input_proc.

Safety

TODO document how to use this function safely.

Object Safety

This trait is not object safe.