Trait coreaudio::audio_unit::render_callback::data::Data

pub trait Data {
    fn does_stream_format_match(&StreamFormat) -> bool;
    unsafe fn from_input_proc_args(num_frames: u32, io_data: *mut AudioBufferList) -> Self;
}

Audio data wrappers specific to the AudioUnit's AudioFormat.

Required Methods

fn does_stream_format_match(&StreamFormat) -> bool

Check whether or not the stream format matches this type of data.

unsafe fn from_input_proc_args(num_frames: u32, io_data: *mut AudioBufferList) -> Self

We must be able to construct Self from arguments given to the input_proc.