Enum coreaudio::audio_unit::types::EffectType

pub enum EffectType {
    PeakLimiter,
    DynamicsProcessor,
    LowPassFilter,
    HighPassFilter,
    BandPassFilter,
    HighShelfFilter,
    LowShelfFilter,
    ParametricEQ,
    Distortion,
    Delay,
    SampleDelay,
    GraphicEQ,
    MultiBandCompressor,
    MatrixReverb,
    Pitch,
    AUFilter,
    NetSend,
    RogerBeep,
    NBandEQ,
}

Effect (digital signal processing) audio unit subtypes for audio units provided by Apple.

Variants
PeakLimiter

An audio unit that enforces an upper dynamic limit on an audio signal.

Available in OS X v10.2 and later.

DynamicsProcessor

An audio unit that provides dynamic compression or expansion.

Available in OS X v10.3 and later.

LowPassFilter

An audio unit that passes frequencies below a specified cutoff frequency and blocks frequencies above that cutoff frequency.

Available in OS X v10.2 and later.

HighPassFilter

An audio unit that passes frequencies above a specified cutoff frequency and blocks frequencies below that cutoff frequency.

Available in OS X v10.2 and later.

BandPassFilter

An audio unit that passes frequencies between specified upper and lower cutoff frequencies, and blocks frequencies outside that band.

Available in OS X v10.2 and later.

HighShelfFilter

An audio unit suitable for implementing a treble control in an audio playback or recording system.

Available in OS X v10.2 and later.

LowShelfFilter

An audio unit suitable for implementing a bass control in an audio playback or recording system.

Available in OS X v10.2 and later.

ParametricEQ

An audio unit that provides a filter whose center frequency, boost/cut level, and Q can be adjusted.

Available in OS X v10.2 and later.

Distortion

An audio unit that provides a distortion effect.

Available in OS X v10.5 and later.

Delay

An audio unit that introduces a time delay to a signal.

Available in OS X v10.2 and later.

SampleDelay

An audio unit that provides a time delay for a specified number of samples.

Available in OS X v10.4 and later.

GraphicEQ

An audio unit that provides a 10- or 31-band graphic equalizer.

Available in OS X v10.2 and later.

MultiBandCompressor

An audio unit that provides four-bands of dynamic compression or expansion.

Available in OS X v10.3 and later.

MatrixReverb

An audio unit that provides a reverberation effect that can be used to simulate a variety of acoustic spaces.

Available in OS X v10.2 and later.

Pitch

An audio unit for modifying the pitch of a signal.

Available in OS X v10.4 and later.

AUFilter

An audio unit that provides a combination of five filters: low-frequency, three mid-frequencies, and high-frequency.

Available in OS X v10.4 and later.

NetSend

An audio unit for use in conjunction with a kAudioUnitSubType_NetReceive audio unit for sending audio across a network or from one application to another.

Available in OS X v10.4 and later.

RogerBeep

An audio unit that detects gaps between segments of speech and fills the gaps with a short tone, simulating the sound of a walkie-talkie communication device.

Available in OS X v10.5 and later.

NBandEQ

A multi-band equalizer with specifiable filter type for each band.

Available in OS X v10.9 and later.